support using builtin interfaces from user types
