+++ aaa/src/common/mod.rs
@@ -1,4 +1,5 @@
 pub mod files;
 pub mod formatting;
+pub mod hash;
 pub mod position;
 pub mod traits;
